Created game loop
The game now has the game loop that cycles though the players turns! This includes asking the player to pass in the device (I wrote a method to do this so use it other methods), rolling the dices, asking the player what they want to do on their turn (move out of room, make guess, etc) based on their current situation (if they are in a room or not), skipping "out" players turns and ending the game if all players are "out". (A player is "out" if they make an incorrect solve attempt). Now all that needs to be added is the makeGuess(), trySolve() and movePlayer() methods and the map display created!

Lastly I did test everything I added with the debugger (and I mean EVERYTHING).