Added state changing code
Added a lot of code to Game for state changing to replace the game loop. This entails ten methods for the ten game states (except GameOver) which are to be called by the controller (GUI). I can't test them currently so if you use them be mindful of bugs. So now you should now be able to make any GUI feature you make call the correct state change code when interacted with! I also made a roll dice button if you want an example of a GUI object.

To make the state machine work there are quite a few new global variables in game that were previously local.

There is a new interface "Location" which square and room implement. It has absolutely nothing in it, it just so rooms and square can be grouped together. This is from methods like canMoveTo().

I also added a few more useful methods like canMoveTo() that GUI should use.

Lastly I added @Deprecated before method I found (there will be more) that we should NOT use anymore that can be deleted (I left them in for reference).

Any Questions ask me!